Removing glib references in ds3_uint64_string_map.h
