+++ src/CMakeLists.txt
@@ -8,7 +8,8 @@ add_library(ds3 SHARED
             ds3_utils.h ds3_utils.c
             ds3_connection.h ds3_connection.c
             ds3_uint64_string_map.h ds3_uint64_string_map.c
-            ds3_response_header_utils.h ds3_response_header_utils.c)
+            ds3_response_header_utils.h ds3_response_header_utils.c
+            ds3_bool.h)
 
 if (WIN32)
     set(CMAKE_BUILD_TYPE Release)
@@ -85,6 +86,7 @@ else(WIN32)
                 "ds3_utils.h"
                 "ds3_uint64_string_map.h"
                 "ds3_response_header_utils.h"
+                "ds3_bool.h"
             DESTINATION
                 "/usr/local/include")
     install(TARGETS ds3 DESTINATION lib)
